+++ common-jei/src/main/java/com/yanny/ali/compatibility/jei/JeiGameplayLoot.java
@@ -1,8 +1,9 @@
 package com.yanny.ali.compatibility.jei;
 
-import com.yanny.ali.api.Rect;
+import com.yanny.ali.api.*;
 import com.yanny.ali.compatibility.common.GameplayLootType;
 import com.yanny.ali.compatibility.common.GenericUtils;
+import com.yanny.ali.plugin.client.widget.LootTableWidget;
 import com.yanny.ali.registries.LootCategory;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
@@ -14,7 +15,7 @@ import net.minecraft.network.chat.Component;
 import org.jetbrains.annotations.NotNull;
 
 public class JeiGameplayLoot extends JeiBaseLoot<GameplayLootType, String> {
-    public JeiGameplayLoot(IGuiHelper guiHelper, RecipeType<GameplayLootType> recipeType, LootCategory<String> lootCategory, Component title, IDrawable icon) {
+    public JeiGameplayLoot(IGuiHelper guiHelper, RecipeType<RecipeHolder<GameplayLootType>> recipeType, LootCategory<String> lootCategory, Component title, IDrawable icon) {
         super(guiHelper, recipeType, lootCategory, title, icon);
     }
 
@@ -25,11 +26,11 @@ public class JeiGameplayLoot extends JeiBaseLoot<GameplayLootType, String> {
     }
 
     @Override
-    public void draw(GameplayLootType recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY) {
+    public void draw(RecipeHolder<GameplayLootType> recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY) {
         super.draw(recipe, recipeSlotsView, guiGraphics, mouseX, mouseY);
-        String key = "ali/loot_table/" + recipe.id().substring(1);
-        Component text = GenericUtils.ellipsis(key, recipe.id(), 9 * 18);
-        Component fullText = Component.translatableWithFallback(key, recipe.id());
+        String key = "ali/loot_table/" + recipe.type().id().substring(1);
+        Component text = GenericUtils.ellipsis(key, recipe.type().id(), 9 * 18);
+        Component fullText = Component.translatableWithFallback(key, recipe.type().id());
         Rect rect = new Rect(0, 0, 9 * 18, 8);
 
         guiGraphics.drawString(Minecraft.getInstance().font, text, 0, 0, 0, false);
@@ -42,5 +43,10 @@ public class JeiGameplayLoot extends JeiBaseLoot<GameplayLootType, String> {
     @Override
     int getYOffset(GameplayLootType recipe) {
         return 10;
+    }
+
+    @Override
+    IWidget getRootWidget(IWidgetUtils utils, IDataNode entry, RelativeRect rect, int maxWidth) {
+        return new LootTableWidget(utils, entry, rect, maxWidth);
     }
 }