+++ common/src/main/java/com/yanny/ali/manager/AliServerRegistry.java
@@ -6,7 +6,9 @@ import com.mojang.serialization.MapCodec;
 import com.yanny.ali.api.*;
 import com.yanny.ali.plugin.common.nodes.LootTableNode;
 import com.yanny.ali.plugin.common.nodes.MissingNode;
+import com.yanny.ali.plugin.common.trades.TradeNode;
 import com.yanny.ali.plugin.server.GenericTooltipUtils;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.advancements.critereon.EntitySubPredicate;
 import net.minecraft.advancements.critereon.ItemSubPredicate;
 import net.minecraft.core.Holder;
@@ -17,6 +19,7 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.npc.VillagerTrades;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -33,6 +36,7 @@ import org.apache.commons.lang3.function.TriFunction;
 import org.apache.logging.log4j.util.TriConsumer;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
+import oshi.util.tuples.Pair;
 
 import java.util.*;
 import java.util.function.BiFunction;
@@ -45,26 +49,35 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
     private final Map<Class<?>, TriFunction<IServerUtils, List<Item>, LootItemFunction, List<Item>>> functionItemCollectorMap = new HashMap<>();
     private final Map<Class<?>, BiFunction<IServerUtils, NumberProvider, RangeValue>> numberConverterMap = new HashMap<>();
     private final Map<Class<?>, EntryFactory<?>> entryFactoryMap = new HashMap<>();
+
     private final Map<Class<?>, BiFunction<IServerUtils, LootItemFunction, ITooltipNode>> functionTooltipMap = new HashMap<>();
     private final Map<Class<?>, BiFunction<IServerUtils, LootItemCondition, ITooltipNode>> conditionTooltipMap = new HashMap<>();
     private final Map<Class<?>, BiFunction<IServerUtils, Ingredient, ITooltipNode>> ingredientTooltipMap = new HashMap<>();
     private final Map<Class<?>, BiFunction<IServerUtils, ItemSubPredicate, ITooltipNode>> itemSubPredicateTooltipMap = new HashMap<>();
     private final Map<MapCodec<?>, BiFunction<IServerUtils, EntitySubPredicate, ITooltipNode>> entitySubPredicateTooltipMap = new HashMap<>();
     private final Map<DataComponentType<?>, BiFunction<IServerUtils, Object, ITooltipNode>> dataComponentTypeTooltipMap = new HashMap<>();
+
     private final Map<Class<?>, TriConsumer<IServerUtils, LootItemCondition, Map<Holder<Enchantment>, Map<Integer, RangeValue>>>> chanceModifierMap = new HashMap<>();
     private final Map<Class<?>, TriConsumer<IServerUtils, LootItemFunction, Map<Holder<Enchantment>, Map<Integer, RangeValue>>>> countModifierMap = new HashMap<>();
     private final Map<Class<?>, TriFunction<IServerUtils, LootItemFunction, ItemStack, ItemStack>> itemStackModifierMap = new HashMap<>();
+
     private final Map<ResourceKey<LootTable>, LootTable> lootTableMap = new HashMap<>();
     private final List<Function<IServerUtils, List<ILootModifier<?>>>> lootModifierGetters = new LinkedList<>();
     private final List<ILootModifier<?>> lootModifierMap = new LinkedList<>();
-    private final ICommonUtils utils;
 
+    private final Map<Class<?>, BiFunction<IServerUtils, VillagerTrades.ItemListing, IDataNode>> itemListingFactoryMap = new HashMap<>();
+    private final Map<Class<?>, BiFunction<IServerUtils, VillagerTrades.ItemListing, Pair<List<Item>, List<Item>>>> tradeItemCollectorMap = new HashMap<>();
+
+    private final Set<Class<?>> missingEntryFactories = new HashSet<>();
     private final Set<Class<?>> missingFunctionTooltips = new HashSet<>();
     private final Set<Class<?>> missingConditionTooltips = new HashSet<>();
     private final Set<Class<?>> missingIngredientTooltips = new HashSet<>();
     private final Set<Class<?>> missingItemSubPredicateTooltips = new HashSet<>();
     private final Set<Class<?>> missingEntitySubPredicateTooltips = new HashSet<>();
     private final Set<Class<?>> missingDataComponentTypeTooltips = new HashSet<>();
+    private final Set<Class<?>> missingItemListingFactories = new HashSet<>();
+
+    private final ICommonUtils utils;
 
     private ServerLevel serverLevel;
     private LootContext lootContext;
@@ -94,12 +107,17 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         lootModifierGetters.clear();
         lootModifierMap.clear();
 
+        itemListingFactoryMap.clear();
+        tradeItemCollectorMap.clear();
+
+        missingEntryFactories.clear();
         missingFunctionTooltips.clear();
         missingConditionTooltips.clear();
         missingIngredientTooltips.clear();
         missingItemSubPredicateTooltips.clear();
         missingEntitySubPredicateTooltips.clear();
         missingDataComponentTypeTooltips.clear();
+        missingItemListingFactories.clear();
     }
 
     public void addLootTable(ResourceKey<LootTable> resourceLocation, LootTable lootTable) {
@@ -204,6 +222,18 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
     }
 
     @Override
+    public <T extends VillagerTrades.ItemListing> void registerItemListing(Class<T> type, BiFunction<IServerUtils, T, IDataNode> tradeFactory) {
+        //noinspection unchecked
+        itemListingFactoryMap.put(type, (u, i) -> tradeFactory.apply(u, (T) i));
+    }
+
+    @Override
+    public <T extends VillagerTrades.ItemListing> void registerItemListingCollector(Class<T> type, BiFunction<IServerUtils, T, Pair<List<Item>, List<Item>>> itemSupplier) {
+        //noinspection unchecked
+        tradeItemCollectorMap.put(type, (u, i) -> itemSupplier.apply(u, (T) i));
+    }
+
+    @Override
     public <T extends LootPoolEntryContainer> List<Item> collectItems(IServerUtils utils, T entry) {
         BiFunction<IServerUtils, LootPoolEntryContainer, List<Item>> itemSupplier = entryItemCollectorMap.get(entry.getClass());
 
@@ -230,7 +260,12 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         //noinspection unchecked
         EntryFactory<T> entryFactory = (EntryFactory<T>) entryFactoryMap.get(type.getClass());
 
-        return Objects.requireNonNullElseGet(entryFactory, () -> (utils1, entry, chance, sumWeight, functions, conditions) -> new MissingNode());
+        if (entryFactory != null) {
+            return entryFactory;
+        } else {
+            missingEntryFactories.add(type.getClass());
+            return (utils1, entry, chance, sumWeight, functions, conditions) ->  new MissingNode();
+        }
     }
 
     @Override
@@ -347,6 +382,31 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
     }
 
     @Override
+    public <T extends VillagerTrades.ItemListing> BiFunction<IServerUtils, T, IDataNode> getItemListingFactory(IServerUtils utils, T entry) {
+        //noinspection unchecked
+        BiFunction<IServerUtils, T, IDataNode> itemListingFactory = (BiFunction<IServerUtils, T, IDataNode>) itemListingFactoryMap.get(entry.getClass());
+
+        if (itemListingFactory != null) {
+            return itemListingFactory;
+        } else {
+            missingItemListingFactories.add(entry.getClass());
+            return (utils1, entry1) ->  new MissingNode();
+        }
+    }
+
+    @Override
+    public <T extends VillagerTrades.ItemListing> Pair<List<Item>, List<Item>> collectItems(IServerUtils utils, T entry) {
+        //noinspection unchecked
+        BiFunction<IServerUtils, T, Pair<List<Item>, List<Item>>> itemCollector = (BiFunction<IServerUtils, T, Pair<List<Item>, List<Item>>>) tradeItemCollectorMap.get(entry.getClass());
+
+        if (itemCollector != null) {
+            return itemCollector.apply(utils, entry);
+        }
+
+        return new Pair<>(Collections.emptyList(), Collections.emptyList());
+    }
+
+    @Override
     public RangeValue convertNumber(IServerUtils utils, @Nullable NumberProvider numberProvider) {
         if (numberProvider != null) {
             BiFunction<IServerUtils, NumberProvider, RangeValue> function = numberConverterMap.get(numberProvider.getClass());
@@ -393,6 +453,10 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         return new LootTableNode(modifiers, this, lootTable);
     }
 
+    public IDataNode parseTrade(Int2ObjectMap<VillagerTrades.ItemListing[]> itemListingMap) {
+        return new TradeNode(this, itemListingMap);
+    }
+
     @Override
     public List<Entity> createEntities(EntityType<?> type, Level level) {
         return utils.createEntities(type, level);
@@ -413,15 +477,18 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         LOGGER.info("Registered {} count modifiers", countModifierMap.size());
         LOGGER.info("Registered {} item stack modifiers", itemStackModifierMap.size());
         LOGGER.info("Registered {} loot modifiers", lootModifierMap.size());
+        LOGGER.info("Registered {} item listing factories", itemListingFactoryMap.size());
     }
 
     public void printRuntimeInfo() {
+        missingEntryFactories.forEach((t) -> LOGGER.warn("Missing entry factory for {}", t.getName()));
         missingFunctionTooltips.forEach((t) -> LOGGER.warn("Missing function tooltip for {}", t.getCanonicalName()));
         missingConditionTooltips.forEach((t) -> LOGGER.warn("Missing condition tooltip for {}", t.getCanonicalName()));
         missingIngredientTooltips.forEach((t) -> LOGGER.warn("Missing ingredient tooltip for {}", t.getCanonicalName()));
         missingItemSubPredicateTooltips.forEach((t) -> LOGGER.warn("Missing item sub predicate tooltip for {}", t.getCanonicalName()));
         missingEntitySubPredicateTooltips.forEach((t) -> LOGGER.warn("Missing entity sub predicate tooltip for {}", t.getCanonicalName()));
         missingDataComponentTypeTooltips.forEach((t) -> LOGGER.warn("Missing data component type tooltip for {}", t.getCanonicalName()));
+        missingItemListingFactories.forEach((t) -> LOGGER.warn("Missing trade item listing for {}", t.getName()));
         LOGGER.info("Prepared {} loot tables", lootTableMap.size());
     }
 }