+++ common/src/main/java/com/yanny/ali/manager/IOnDoneListener.java
@@ -2,11 +2,12 @@ package com.yanny.ali.manager;
 
 import com.yanny.ali.api.IDataNode;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.storage.loot.LootTable;
 
 import java.util.Map;
 
 @FunctionalInterface
 public interface IOnDoneListener {
-    void onDone(Map<ResourceKey<LootTable>, IDataNode> lootData);
+    void onDone(Map<ResourceKey<LootTable>, IDataNode> lootData, Map<ResourceLocation, IDataNode> tradeData);
 }