+++ common/src/main/java/com/yanny/ali/plugin/client/widget/ItemWidget.java
@@ -3,11 +3,10 @@ package com.yanny.ali.plugin.client.widget;
 import com.yanny.ali.api.*;
 import com.yanny.ali.plugin.common.nodes.ItemNode;
 
-public class ItemWidget extends IWidget {
+public class ItemWidget implements IWidget {
     private final RelativeRect bounds;
 
     public ItemWidget(IWidgetUtils utils, IDataNode entry, RelativeRect rect, int maxWidth) {
-        super(entry.getId());
         ItemNode node = (ItemNode) entry;
         utils.addSlotWidget(node.getModifiedItem(), node, rect);
         bounds = rect;