+++ fabric/src/main/java/com/yanny/ali/network/Client.java
@@ -6,6 +6,9 @@ public class Client extends AbstractClient {
     public void onLootInfo(SyncLootTableMessage message, ClientPlayNetworking.Context context) {
         super.onLootInfo(message);
     }
+    public void onTradeInfo(Minecraft client, ClientPacketListener handler, FriendlyByteBuf buf, PacketSender responseSender) {
+        super.onTradeInfo(new SyncTradeMessage(buf));
+    }
 
     public void onClear(ClearMessage message, ClientPlayNetworking.Context context) {
         super.onClear(message);